+++ pkg/lint/rules/should_disable_additional_properties_test.go
@@ -34,7 +34,7 @@ func TestShouldDisableAdditionalProperties(t *testing.T) {
 				t.Fatalf("Unexpected parsing error in test case '%s': %s", tc.name, err)
 			}
 
-			ruleResults := []string{}
+			ruleResults := []lint.RuleViolation{}
 			for _, rule := range tc.rules {
 				ruleResults = append(ruleResults, rule.Verify(schema).Violations...)
 			}