+++ CHANGELOG.md
@@ -8,10 +8,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add rule to check whether root-level specifies properties of the common schema structure.
+
+## [1.0.1] - 2023-03-22
+
+- Fix module path.
+
+## [1.0.0] - 2023-03-22
+
+- Fix update action workflow trigger.
+- Add a reusable composite GitHub action that calls `schemalint verify`. 
+
+## [0.10.0] - 2023-03-07
+
+- Add possibility to exclude locations from rule set validation.
+
+## [0.9.0] - 2023-03-02
+
+- Check if normalization is already applied when calling normalize with output argument.
 - Add rules to forbid infinite recursion and recursion-related keywords.
 - Add possibility to output version with `schemalint -v` or `schemalint --version`.
 - Add rule to check whether logical constructs (if, then & else) are not used.
 - Add check that schemas only use `anyOf` and `oneOf` for specific purposes.
+- Add rule to check that arrays only specify one type for their items.
 
 ## [0.8.0] - 2023-02-24
 
@@ -73,7 +91,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added first basic linting.
 
-[Unreleased]: https://github.com/giantswarm/schemalint/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/schemalint/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/giantswarm/schemalint/compare/v1.0.0...v1.0.1
+[1.0.0]: https://github.com/giantswarm/schemalint/compare/v0.10.0...v1.0.0
+[0.10.0]: https://github.com/giantswarm/schemalint/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/giantswarm/schemalint/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/giantswarm/schemalint/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/giantswarm/schemalint/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/schemalint/compare/v0.5.0...v0.6.0